requirements.txt
Django==3.2
django-ckeditor==6.0.0